Move Cursor tests to own test classes
